backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
pandas==2.1.4
numpy==1.24.3
python-multipart==0.0.6
cors==1.0.1
fastapi-cors==0.0.6
